+++ Hazelcast.Net/Hazelcast.Client.Request.Concurrent.Lock/LockRequest.cs
@@ -5,11 +5,13 @@ namespace Hazelcast.Client.Request.Concurrent.Lock
 {
     internal sealed class LockRequest : AbstractLockRequest
     {
-        public LockRequest(IData key, int threadId) : base(key, threadId)
+        public LockRequest(IData key, long threadId)
+            : base(key, threadId)
         {
         }
 
-        public LockRequest(IData key, int threadId, long ttl, long timeout) : base(key, threadId, ttl, timeout)
+        public LockRequest(IData key, long threadId, long ttl, long timeout)
+            : base(key, threadId, ttl, timeout)
         {
         }
